pyproject.toml
[project]
name = "django-test"
version = "0.1.0"
description = "Some description"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "django-silk>=5.4.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.4",
]
prod = [
    "gunicorn>=23.0.0",
]
